setup-user: restore portability
Changes in commit 5129e97285 (setup-user: switch to git-var to check if
username and e-mail are set, 2019-11-24, v3.17.0-rc1~379^2) now require
the `rev` and `cut` tools to be available.  Revise the logic to use that
approach only when they are available, and otherwise fall back to the
previous more portable approach.

Co-Author: Brad King <brad.king@kitware.com>