Tutorial: Restore MakeTable.cxx in step 6
In commit c754a3d4b7 (Tutorial: Remove MakeTable.cxx from Steps 5 and 6,
2020-04-23) it was incorrect to remove the file from step 6.  The
instructions for that step show the addition of a reference to it from
the `CMakeLists.txt` file.  Each step shows the addition of content to
lead to the next step, so removing the file from step 6 was an
off-by-one error.

Issue: #20618