VisualStudio: move PCH rules to projects when possible.
This dramatically helps reduce the size of the solution files
when PCH is enabled, since 2 entries per source file are removed.

This also corrects a subtle issue where when UNITY + PCH was enabled,
the PCH would not be used if a user explicitly tried to compile
a source file from outside the unity group. This is possible via
the compile source option in the Visual Studio GUI.