Merge from upstream. (#3)
* clang-tidy: ignore `misc-no-recursion`

CMake has intentional recursion in its codebase, so avoid warning about
occurrences.

* clang-tidy: ignore new warnings from newer versions

These warnings can be fixed in the future, but are ignored for now.

* Tests/FindBoost/TestPython: support finding 3.8 and 3.9

* ci: install gmock in the Fedora 31 image

The gmock libraries are referenced by GTest's CMake package files.

While at it, also explicitly install gettext.  It is installed already,
but make it explicit for future testing of the FindGettext module.

(cherry picked from commit 585cc7c9301ff474a40d7702ce429b93f474d44f)

* ci: update to Fedora 33 for Linux builds

* gitlab-ci: update Linux CI to use Fedora 33

* gitlab-ci: use Fedora 33 to build sources

* MSVC: Factor out initialization of /GR flag

* MSVC: Do not add /GR to CMAKE_CXX_FLAGS by default

The `/GR` flag has been on by default since MSVC cl 14.0 from VS 2005.
Remove it from the default flags to make it easier for projects to pass
`/GR-` themselves to turn it off.

Projects may be using string processing to replace `/GR` with another
flag, so we cannot simply drop it.  Add a policy to drop it in a
compatible way.

Fixes: #21428

* CMake Nightly Date Stamp

* Xcode: Clean library paths to avoid linker duplicate symbol definitions

* CMake Nightly Date Stamp

* Makefiles: Explicitly tell 'make' tool what Makefile name to use

CMake did not specify the filename of the Makefile generated by it.
Due to GNU make precedence rules this means that the presence of a
GNUmakefile or makefile would take precedence over the generated
Makefile.

This is only relevant for in-source builds and only whenever an
alternative makefile by the above mentioned names exists.

This patch adds the (seemingly universal) `-f` switch and the
(hardcoded) filename (it is now hardcoded separately in these two
files):

- cmLocalUnixMakefileGenerator3.cxx
- cmGlobalUnixMakefileGenerator3.cxx

Fixes: #21418

* Tests: Add RunCMake.Make case for conflicting Makefile names

Issue: #21418

* Ninja: Exclude unused dyndep features during CMake bootstrap

* PCH: Do not mark PCH-header as generated

The PCH header file itself is written by CMake and is not generated by
part of the buildsystem.  Therefore the `GENERATED` property is not
appropriate to set because the file is always present before the build
starts.  We already do not mark the PCH source file as `GENERATED`
either.

This is a preparation for solving #18399, which will no longer allow to
unset the `GENERATED` property from a source-file once it was set.

Fixes: #21437
Signed-off-by: Deniz Bahadir <dbahadir@benocs.com>

* CPack/DEB: Do not crash when asked for debug symbols when there are none

Fixes: #21356

* CMake Nightly Date Stamp

* Suppress MSVC 14.28 C5105 warning triggered by Windows SDK

Issue: #21446

* cmScanDepFormat: add reader and writer for the format from P1689R2

This format is currently subject to change, but is not too far from the
end goal.

Some bits are currently unimplemented (see TODO comments).

* cmGlobalNinjaGenerator: use P1689 dependency file format for Fortran

The module dependency specification format described in the C++
JTC1/SC22/WG21 paper [1] is also suitable for use by Fortran.

[1] http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2020/p1689r2.html

* cmake: Fix processing of -Wno-error= flags

Fix two bugs that happened to cancel each other out for cases covered
by our test suite.  Add a test case that distinguishes them.

* gitlab-ci: Update Windows builds to MSVC 19.28

* CTest: Rename CudaMemcheck to CudaSanitizer

* CMake Nightly Date Stamp

* CMake 3.18.5

* CMake 3.19.0

* scandep: update format to P1689 R3

The `work-directory` field has been moved into individual rules since
each one may correspond to a different compiler invocation and therefore
a different work directory.

Support for arbitrary binary paths has been removed for simplicity.
We now assume paths have a valid UTF-8 round-trip representation.

* ci: update to use CMake 3.19.0

* gitlab-ci: update macOS jobs to use Xcode 12.0

* cmake.el: enable view-mode when lookup

* CMake Nightly Date Stamp

* cmake: redesign command-line argument parsing

Make handling more consistent:

   "-S" -> invalid
   "-S/path/to/source" -> valid
   "-S /path/to/source" -> valid
   "-S=/path/to/source" -> now valid

   "-D" -> invalid
   "-DStr" -> valid
   "-D Str" -> valid
   "-D=Str" -> now valid

   "--log-level=" -> invalid
   "--log-level" -> invalid
   "--log-level=2" -> valid
   "--log-level 2" -> now valid

* cmake: Simplify -W recognition of no- and error= prefixes

* Revert "specify language flag when source LANGUAGE property is set"

Revert commit 74b1c9fc8e (Explicitly specify language flag when source
LANGUAGE property is set, 2020-06-01, v3.19.0-rc1~722^2) and the lookup
tables from its two immediate ancestors.  The purpose of that change was
to convert an explicit `LANGUAGE` source file property into an explicit
language specification compiler flag like `-x c`.  This seems reasonable
since the property is documented as meaning "indicate what programming
language the source file is".  It is also needed to help compilers deal
with non-standard source file extensions they don't recognize.

However, some projects have been setting `LANGUAGE C` on `.S` assembler
source files to mean "use the C compiler".  Passing `-x c` for them
breaks the build because the `.S` sources are not written in C.  These
projects should be updated to use `enable_language(ASM)`, for which
CMake often chooses the C compiler as the assembler when using
toolchains that support it (which would have to be the case for projects
using the approach).

Revert the change for now to preserve the old behavior for such projects.
We can re-introduce it with a policy in a future version of CMake.

Fixes: #21469
Issue: #14516, #20716

* CMake Nightly Date Stamp

Co-authored-by: Ben Boeckel <ben.boeckel@kitware.com>
Co-authored-by: Brad King <brad.king@kitware.com>
Co-authored-by: Kitware Robot <kwrobot@kitware.com>
Co-authored-by: Gusts Kaksis <gusts.kaksis@sonarworks.com>
Co-authored-by: Oliver Schneider <oliver@assarbad.net>
Co-authored-by: Deniz Bahadir <dbahadir@benocs.com>
Co-authored-by: Andrew Fuller <afuller@teradici.com>
Co-authored-by: Craig Scott <craig.scott@crascit.com>
Co-authored-by: Tobias Ribizel <ribizel@kit.edu>
Co-authored-by: condy <condy0919@gmail.com>
Co-authored-by: Robert Maynard <robert.maynard@kitware.com>