Tests: Avoid unnecessary use of toolchain in RunCMake.CMP0118
The semantics of the `GENERATED` property do not depend on the kind of
target to which the sources are attached.  Convert the tests to use
custom targets instead of executables so that we do not have to enable
the CXX language and initialize support of its toolchain for every case.
This significantly reduces the time the test takes to run.