Tests: Remove outdated CMakeTestMultipleConfigures test
In practice this test has not helped us find any problems AFAIK.  Tests
like RunCMake.BuildDepends and other RunCMake tests cover much of what
this test was meant to do.  Since CMakeTestMultipleConfigures takes a
long time to run on some generator/platform combinations, it is not
worth what it covers.  Remove the test.