Tests: Generate simpler ExternalProject sample Git repositories
The `gitrepo.tgz` repository contains a project that enables both
C and CXX languages and builds source code.  This take a long time
to repeat over and over, particularly on Visual Studio and Xcode
generators.  Add and use a script to generate a new test repository
that has a much simpler project.

Issue: #20733