AUTOUIC: Collect ui header files for Ninja generator
The '<user target>_autogen/timestamp' target supposed to
generate ui header files using the 'uic'. Ninja must have
information about these header files as a result of generating.

The fix collects .ui files of the user target and generates
a list of the ui headers that need to be added to the generating
results of the '<user target>_autogen/timestamp' target.

The case when the .ui files are not specified and collected by
AUTOUIC from the include directives of the project source files
is not covered in this patch.

Fixes: #16776